+++ database/__init__.py
@@ -1,3 +1,3 @@
 from .postgres import Database
-from .async_postgres import AsyncDatabase
 
+__all__ = ('Database',)